% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{abgd_delim}
\alias{abgd_delim}
\title{Perform Species Delimitation with ABGD Result}
\usage{
abgd_delim(output_folder, indices, label = NULL)
}
\arguments{
\item{output_folder}{Output folder of the ABGD result.}

\item{indices}{A numerical vector indicating the indices of the partitions to be evaluated.}

\item{label}{A character vector. Name to added to the column name of the output data frame. The length must match with \code{indices}.}
}
\value{
A data frame of the ABGD result.
}
\description{
Perform Species Delimitation with ABGD Result
}
\seealso{
\code{\link[=abgd_execute]{abgd_execute()}}

\code{\link[=abgd_read_result]{abgd_read_result()}}

\code{\link[=abgd_read_result_all]{abgd_read_result_all()}}

\code{\link[=abgd_summary]{abgd_summary()}}
}
