% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{abgd_execute}
\alias{abgd_execute}
\title{Perform Automatic Barcoding Gap Detection (ABGD)}
\usage{
abgd_execute(
  alignment,
  exe,
  output_folder = NULL,
  a = T,
  p = 0.001,
  P = 0.1,
  n = 20,
  d = 3,
  X = 1.5,
  args = NULL
)
}
\arguments{
\item{alignment}{Path to the FASTA file.}

\item{exe}{Path to the ABGD executable.}

\item{output_folder}{Path to output folder.}

\item{p}{Minimal a priori value (Pmin, default to 0.001).}

\item{P}{Maximal a priori value (Pmax, default to 0.1).}

\item{n}{Number of steps in [Pmin, Pmax] (default to 20)}

\item{d}{Distance to use (0: Kimura-2P, 1: Jukes-Cantor, 2: Tamura-Nei, 3:simple distance; default to 3).}

\item{X}{Mininmum Slope Increase (default to 1.5)}
}
\value{
Path to the output folder.

The output folder where ABGD CLI program will export the result
}
\description{
Perform Automatic Barcoding Gap Detection (ABGD)
}
\seealso{
\code{\link[=abgd_read_result]{abgd_read_result()}}

\code{\link[=abgd_read_result_all]{abgd_read_result_all()}}

\code{\link[=abgd_summary]{abgd_summary()}}

\code{\link[=abgd_delim]{abgd_delim()}}
}
