% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{abgd_read_result_all}
\alias{abgd_read_result_all}
\title{Read All ABGD Partition Results}
\usage{
abgd_read_result_all(output_folder)
}
\value{
A named list of summary for all partitions
}
\description{
A bulk loading version of \code{abgd_read_result}.
}
\seealso{
\code{\link[=abgd_execute]{abgd_execute()}}

\code{\link[=abgd_read_result]{abgd_read_result()}}

\code{\link[=abgd_summary]{abgd_summary()}}

\code{\link[=abgd_delim]{abgd_delim()}}
}
