% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{abgd_summary}
\alias{abgd_summary}
\title{Generate Summary for ABGD Result}
\usage{
abgd_summary(output_folder, silent = F)
}
\arguments{
\item{output_folder}{Path to the ABGD output folder.}

\item{silent}{Logical. Whether show plot or not.}
}
\value{
A list of length 2.
\describe{
\item{\code{plot}}{A plot showing number of groups fround with each partition will be shown.}
\item{\code{summary}}{A data.frame of number of groups and the prior intraspecific divergence of each partition will be returned.}

A named list. \describe{
\item{\code{plot}}{A ggplot object, the summary plot}
\item{\code{summary}}{A data frame of the partition summary}
}
}
\description{
Generate Summary for ABGD Result
}
\seealso{
\code{\link[=abgd_execute]{abgd_execute()}}

\code{\link[=abgd_read_result]{abgd_read_result()}}

\code{\link[=abgd_read_result_all]{abgd_read_result_all()}}

\code{\link[=abgd_delim]{abgd_delim()}}
}
