% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{bathy.basemap}
\alias{bathy.basemap}
\title{Create a Bathymetric and Topographic Basemap}
\usage{
bathy.basemap(
  boundaries = list(0, -100, 75, 20),
  vertical.tiles = 5,
  horizontal.tiles = 5,
  checkerboard = T,
  bathy.clr = T,
  land = T,
  crs = 4326,
  resolution = 5,
  z = 3,
  cache = T
)
}
\arguments{
\item{boundaries}{A list of boundaries (min.lat, min.long, max.lat, max.long). Default values are suitable for North Atlantic.}

\item{vertical.tiles}{Number of vertical checkerboard tiles. Will be ignored if \code{checkerboard} if False.}

\item{horizontal.tiles}{Number of horizontal checkerboard tiles. Will be ignored if \code{checkerboard} if False.}

\item{checkerboard}{Logical. Add checkerboard scales?}

\item{bathy.clr}{Logical. Add bathymetry colors?}

\item{land}{Logical. Add land elevation colors?}

\item{crs}{The coordinate reference system.}

\item{resolution}{Resolution for \code{getNOAA.bathy}.}

\item{z}{Zoom level for \code{elevatr::get_elev_raster}.}
}
\description{
(Add your description here...)
}
