% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{beast2_write_mrca}
\alias{beast2_write_mrca}
\title{Write MRCA Prior into Alignment Nexus File for BEAST2}
\usage{
beast2_write_mrca(nexus, monophyly_groups)
}
\arguments{
\item{nexus}{A path to the alignment nexus file.}

\item{monophyly_groups}{A list defining monophyly groups. Each element of this list defines one monophyly group and must be a list itself containing the following two components:
\describe{
\item{\code{name}}{A character string (length 1) specifying the group's name.}
\item{\code{labels}}{A character vector of taxa belonging to that group, this should be identical to the ones in the alginment.}}}
}
\description{
BEAST2 produces full bifurcation ultrametric tree required to perform GMYC. MRCA Prior enforces certain taxa to be monophyly in the resulted tree, which serves as a method of defining the outgroups. The MRCA Prior information could be written in a nexus block and recognized in BEAUti and BEAST2.
}
