% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_clustering.R
\name{cc}
\alias{cc}
\title{Consensus Clustering}
\usage{
cc(
  source_matrix,
  k,
  reps = 500,
  subset = 0.8,
  base_method = "hclust",
  base_call = NULL,
  dist_method = "cor_pearson",
  dist_call = NULL,
  seed = NULL,
  verbose = T,
  ...
)
}
\arguments{
\item{source_matrix}{A matrix. Multivariate dataset on which to perform the CC.}

\item{k}{Numeric. Hyperparameter. Number of clusters to produce. Will be passed to \code{\link[=cutree]{cutree()}} and if \code{base_method = "kmean"}, also to \code{\link[=kmeans]{kmeans()}}.}

\item{reps}{Number of subsampling to build consensus matrix.}

\item{subset}{Proportion of source_matrix to subsample in each replication.}

\item{base_method}{Base clustering method. Either \code{"hclust"} for hierarchical clustering (\code{hclust(method = "ward.D2")}) or \code{"kmeans"} for k-means clustering (\code{kmeans()}). Will be ignored if \code{base_call} is provided.}

\item{base_call}{A function call for base clustering method, available variables are the subsampled matrix \code{m}, calculated distance matrix \code{dist.matrix}, and everything passed to \code{...}.}

\item{dist_method}{Distance calculation method shortcut to generate the calls. Will be ignored if \code{base_method = "kmeans"} or if \code{dist_call} is provided. Accept \code{"euclidean"}, \code{"manhattan"}, \code{"cor_pearson"}, and \code{"cor_spearman"}.}

\item{dist_call}{A function call for calculating distance that is passed to base clustering method. Available variables are the subsampled matrix \code{m}, and everything passed to \code{...}.}

\item{seed}{Numeric vector of length equal to or greater than \code{reps}. The seed will be sequentially applied to each subsampling. Passing the \code{seed} slot product here will ensure reproducibility of existing results.}

\item{...}{Additional arguments for custom function calls. Either for \code{base_call} or \code{dist_call}. You need to add \code{...} in your function call to receive the arguments.}
}
\value{
A S7 object of the class \code{cc}, with following properties:
\describe{
\item{\code{k}}{Number of clusters. \code{k} used for cc.}
\item{\code{reps}}{Number of iterations}
\item{\code{subset}}{Subset proportion of the original matrix at each subsampling}
\item{\code{dist_method}}{Method shortcut used for distance calculation}
\item{\code{dist_call}}{Actual call for distance calculation}
\item{\code{seed}}{Seeds used for each subsampling}
\item{\code{base_method}}{Method shortcut used for base clustering}
\item{\code{base_call}}{Actual call for base clustering}
\item{\code{summary_call}}{Actual call for summarizing clustering}
\item{\code{sample_name}}{Sample names extracted from the \code{source_matrix}}
\item{\code{assignment}}{Clustering result, cluster assignment of each sample}
\item{\code{silhouette_width_avg}}{Average silhouette width}
\item{\code{source_matrix}}{Source matrix}
\item{\code{cc_matrix}}{Generated consensus matrix}
\item{\code{M}}{A matrix composed by number of times each sample pair clustered together}
\item{\code{I}}{A matrix composed by number of times each samples are subsampled together}
\item{\code{heatmap}}{A \link[ggplot2:ggplot]{ggplot2::ggplot} object, heatmap of the result}
\item{\code{silhouette_plot}}{A \link[ggplot2:ggplot]{ggplot2::ggplot} object, silhouette width plot}
\item{\code{summary_clustering}}{Final summarizing clustering result}
}
}
\description{
Perform consensus clustering (CC) with any base method.
}
\seealso{
\code{\link[=ccmk]{ccmk()}}

\code{\link[=consensus_clustering_summarize]{consensus_clustering_summarize()}}
}
