% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_clustering.R
\name{cc_ensemble}
\alias{cc_ensemble}
\alias{cc_ensemble_cspa}
\title{Ensemble Consensus Clustering}
\usage{
cc_ensemble_cspa(
  cc_list = list(),
  weights = NULL,
  k = NULL,
  method = "mean",
  ensemble_call = expr(hclust(as.dist(1 - matrix), method = "ward.D2")),
  ...
)
}
\arguments{
\item{cc_list}{A list of \code{cc} objects. Ideally, these are the "best k" results from different methods (e.g., list(pearson_k5, spearman_k6)).}

\item{weights}{Numeric vector. Weights for each method. If NULL, equal weights are used.}

\item{k}{Numeric. The number of clusters to cut the final ensemble into. If NULL, it suggests a k based on the ensemble structure (optional extension).}

\item{method}{Character. How to combine the matrices. "mean" (arithmetic mean) or "max" (takes the max value, useful if you want to link samples if \emph{any} method finds them similar).}

\item{ensemble_call}{A function call for the final clustering of assembled consensus matrix. Take \code{matrix} as input.}

\item{...}{Additional arguments passed to \code{\link[=consensus_clustering_summarize]{consensus_clustering_summarize()}}, such as \code{heat_name} or \code{plot_labs}.}
}
\value{
A S7 object of the class \code{cc_ensemble}, with following properties:
\describe{
\item{\code{cc_list}}{A list of \link{cc} class object used for assembly}
\item{\code{weights}}{A numeric vector representing weight of each cluster in assembly}
\item{\code{weight.info}}{Additional information about weighting}
\item{\code{k}}{Number of cluster (\code{k}) used for final assembly clustering}
\item{\code{type}}{Type of assembly approach}
\item{\code{method}}{Method of assembly}
\item{\code{ensemble_call}}{Actual function call of the final clustering}
\item{\code{ensemble_matrix}}{A consensus matrix produced by assembling \link{cc} in \code{cc_list}}
\item{\code{source_matrix}}{The original multivariate matrix used to produce all \link{cc} object}
\item{\code{ensemble_clustering}}{Final clustering result}
\item{\code{assignment}}{Cluster assignment of each sample}
\item{\code{heatmap}}{A \link{ggplot} object, the heatmap of the clustering result}
}
}
\description{
Combines multiple consensus clustering objects (cc) into a single ensemble result.

\code{cc_ensemble_cspa} uses a CSPA-like approach.
}
\seealso{
\code{\link[=cc]{cc()}}
}
