% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_clustering.R
\name{cc_eval}
\alias{cc_eval}
\title{Evaluation of a Consensus Clustering or their Ensemble Run}
\usage{
cc_eval(mat, assignment)
}
\description{
Evaluation of a Consensus Clustering or their Ensemble Run
}
