% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/consensus_clustering.R
\name{ccmk}
\alias{ccmk}
\title{Perform Consensus Clustering for a Range of K}
\usage{
ccmk(source_matrix, kmin, kmax, ...)
}
\arguments{
\item{source_matrix}{Multivariate matrix to perform \code{\link[=cc]{cc()}} on.}

\item{kmin}{Integer. Minimum \code{k} of the range.}

\item{kmax}{Integer. Maximum \code{k} of the range.}

\item{...}{Further argument passed to \code{\link[=cc]{cc()}}.}
}
\value{
A S7 object of the class \code{ccmk}.
}
\description{
Perform Consensus Clustering for a Range of K
}
\seealso{
\code{\link[=cc]{cc()}}
}
