% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{concatenate_alignment}
\alias{concatenate_alignment}
\title{Concatenate Multiple Alignment into Multi-locus Alignment}
\usage{
concatenate_alignment(
  fasta_path,
  names.alignments,
  sequence.map,
  identifier,
  accs_col,
  include_strategy = "all",
  include_strategy_loci,
  output_folder,
  run_name = NULL,
  exclude_options = NULL
)
}
\arguments{
\item{fasta_path}{A character vector of path to the fasta file of each individual alignment. The order of the path indicates their order in the alignment.}

\item{sequence.map}{A data frame to map sequences to individuals. Must have columns specified in \code{identifier} and \code{accs_col}}

\item{identifier}{Character. Name of the column in \code{sequence.map} corresponds to the organism's ID.}

\item{accs_col}{A character vector. Names of the column in \code{sequence.map} where accession numbers or sequence names of each organism are stored. Sequence names in the alignment of \code{fasta_path} will be matched against these.}

\item{include_strategy}{Either to only include taxa that has \code{all} sequences in the loci specified in \code{include_strategy_loci}, or taxa that has sequence in \code{any} of the loci.}

\item{include_strategy_loci}{Names of the loci used to check whether a taxa should be included or not.}

\item{output_folder}{A path to the output folder.}

\item{exclude_options}{A list of arguments passed to \code{\link[=exclude_by_meta]{exclude_by_meta()}}, except the first argument. Will used to exclude sequences. A list of the argument list can be provided to perform multiple steps of filtering.}

\item{names.alignment}{A character vector. Name of the loci. Must have the same length as \code{fasta_path}.}
}
\value{
Concatenated alignment will be written into a fasta file. The function returns a named list with information of the concatenated alignment: \describe{
\item{\code{alignment}}{The acutal concatenated alignment object.}
\item{\code{concat_fasta_path}}{Path to the concatenated alignment file}
\item{\code{loci_fasta_path}}{Path to the fasta file of re-extracted alignment of each locus}
\item{\code{taxa}}{Names of the taxa included in the alignment}
\item{\code{loci_partition}}{A data frame. Storing the names of the loci and their length}
}
}
\description{
Take individual alignment in fasta file, pair them based on a sequence mapping table, perform sequence filtering based on metadata (\code{\link[=exclude_by_meta]{exclude_by_meta()}}), glue the paired sequences into concatenated sequence, construct concatenated alignment based on their sequence availability at certain loci.
}
\seealso{
\code{\link[=exclude_by_meta]{exclude_by_meta()}}

\code{\link[=split_concat_alignment]{split_concat_alignment()}}
}
