% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{consensus_clustering}
\alias{consensus_clustering}
\title{Perform Consensus Clustering}
\usage{
consensus_clustering(
  mv.data,
  k,
  reps = 500,
  subset = 0.8,
  method = "hclust",
  seed = NULL
)
}
\description{
Perform Consensus Clustering
}
