% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{consensus_clustering_mk}
\alias{consensus_clustering_mk}
\title{Perform Consensus Clustering for a Range of K}
\usage{
consensus_clustering_mk(
  mv.matrix,
  kmin,
  kmax,
  summary.call = expr(hclust(as.dist(1 - matrix), method = "ward.D2"))
)
}
\description{
Perform Consensus Clustering for a Range of K
}
