% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{consensus_clustering_summarize}
\alias{consensus_clustering_summarize}
\title{Summarize Consensus Clustering Result}
\usage{
consensus_clustering_summarize(
  matrix,
  k,
  summary.call = expr(hclust(as.dist(1 - matrix), method = "ward.D2")),
  heat_name = "Consensus index",
  title = expr(paste0("Consensus clustering (k = ", k, ")")),
  plot_labs = list(labs(subtitle = "Method: hierarchical clustering (ward.D2)")),
  heat_lab = c("Least similar", "Most similar")
)
}
\arguments{
\item{matrix}{Consensus matrix produced by \code{\link[=consensus_clustering]{consensus_clustering()}}.}

\item{k}{Numerical. Number of groups (k) to be passed to \code{summary.call} and \code{\link[stats:cutree]{stats::cutree()}}.}

\item{summary.call}{Final clustering call to group samples according to \code{matrix}.}

\item{heat_name}{Character. Title of the legend for the tile color.}

\item{title}{Character or an expression. Title of the heatmap. Will be passed to \code{\link[=gg_heatmap_hclust]{gg_heatmap_hclust()}}.}

\item{plot_labs}{\code{\link[ggplot2:labs]{ggplot2::labs()}} call wrapped in a list. Will be passed to \code{\link[=gg_heatmap_hclust]{gg_heatmap_hclust()}}.}

\item{heat_lab}{A character vector in length of two. Labels for the min and max value of the color tile legend. Will be passed to \code{\link[=gg_heatmap_hclust]{gg_heatmap_hclust()}}.}
}
\value{
A list of summary result.
\describe{
\item{\code{k}}{The number of clusters.}
\item{\code{matrix}}{The consensus matrix input.}
\item{\code{fc}}{Final clustering result, an \code{\link[stats:hclust]{hclust}} object.}
\item{\code{cl}}{A data frame containing the cluster assignments (Clustering result).}
\item{\code{heatmap}}{The consensus clustering heatmap, a \code{ggplot} object.}
}
}
\description{
WIP, plug in evaluation to generate supporting values.
}
\seealso{
\code{\link[=gg_heatmap_hclust]{gg_heatmap_hclust()}}

\code{\link[=consensus_clustering]{consensus_clustering()}}
}
