% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{convert_nexus2fasta}
\alias{convert_nexus2fasta}
\title{Convert NEXUS alignment to FASTA alignment}
\usage{
convert_nexus2fasta(
  nexus_path,
  fasta_path = NULL,
  safe_name = F,
  remove_outgroup = NULL
)
}
\arguments{
\item{nexus_path}{Path to NEXUS file.}

\item{fasta_path}{Path to FASTA file.}

\item{safe_name}{logical. Whether to modify the name to remove "_".}

\item{remove_outgroup}{Vector of sequence labels to remove. Default to \code{NULL} without removing any sequence. This happens before modifying sequence labels if \code{safe_name} is set to TRUE.}
}
\value{
The FASTA file will be written in the same directory. The path will be returned.
}
\description{
Convert NEXUS alignment to FASTA alignment
}
