% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_config}
\alias{create_config}
\title{Create Configuration File}
\usage{
create_config()
}
\description{
Several functions in this package requires user-specific configuration to work. Setting up a configuration file could reduce repetitive manual input in workflow.
}
