% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{date2assets}
\alias{date2assets}
\title{Generate path to given date's asset folder in log (is this really necessary?)}
\usage{
date2assets(date)
}
\description{
Generate path to given date's asset folder in log (is this really necessary?)
}
