% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_pull}
\alias{db_pull}
\title{Pull Table}
\usage{
db_pull(table, cleaned = T, formatting = T, format = NULL)
}
\arguments{
\item{table}{Table name. Type "show_table" to display all available table names in database.}

\item{cleaned}{Clean table, rows marked in \code{del} column will be dropped.}

\item{formatting}{Format table. Utilize table-specific formatting function to format the table if available. Formatting functions are always named under the rule \code{format_} + table name.}

\item{format}{Argument to pass over to formatting functions. If available different formatting could be chosen.}
}
\description{
Pull Table
}
