% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/database.R
\name{db_sign}
\alias{db_sign}
\title{Generate Signature for Database Change}
\usage{
db_sign(
  table,
  type,
  msg,
  request = request.id,
  con = dbConnect(RSQLite::SQLite(), "data/database/HPLNCdb.sqlite"),
  time = format(Sys.time(), "\%Y-\%m-\%d \%H:\%M:\%S \%Z"),
  signature.table = "signature"
)
}
\arguments{
\item{table}{Table name subjected to change}

\item{type}{Change type}

\item{request}{Request ID}

\item{con}{Connection to database as produced by \code{DBI::dbConnect()}}

\item{time}{Time of change. Default to current time.}

\item{signature.table}{Table name to write signature.}
}
\description{
Generate Signature for Database Change
}
