% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DBchange.R
\name{DBmanual_write}
\alias{DBmanual_write}
\title{Write Table}
\usage{
DBmanual_write(tbl, table, msg, request.id)
}
\arguments{
\item{tbl}{Table object to write.}

\item{table}{Table name.}

\item{msg}{Change message.}

\item{request.id}{Request ID associated to this change.}
}
\description{
Write Table
}
