% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{dec_format}
\alias{dec_format}
\title{Format DD Coordinates with Letters.}
\usage{
dec_format(dec)
}
\arguments{
\item{dec}{DD coordinates. This is a vectorized function, which means you can either pass one character that is latitude or longitude coordinates, or you can pass a vector of those. But not both latitude and longitude at the same time, it only handles one vector at a time.}
}
\description{
Format decimal degree coordinates notated with letters to clean numerical coordinates. West and south with minus.
}
\author{
Zhehao Hu
}
