% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{dms_to_decimal}
\alias{dms_to_decimal}
\title{Convert DMS or DDM Coordinates to DD Coordinates.}
\usage{
dms_to_decimal(dms)
}
\arguments{
\item{dms}{DMS or DDM coordinates. This is a vectorized function, which means you can either pass one character that is latitude or longitude coordinates, or you can pass a vector of those. But not both latitude and longitude at the same time, it only handles one vector at a time.}
}
\value{
Converted decimal degree coordinates, ready to use.
}
\description{
Also handles various notations and direction letter just as they are written in the cruise reports, which is never ready to use.
}
\author{
Zhehao Hu
}
