% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/consensus_clustering.R
\name{evaluate}
\alias{evaluate}
\title{Evaluate Analysis Quality}
\usage{
evaluate(x, ...)

evaluate(cc)

evaluate(cc_ensemble)
}
\description{
A S7 generics to print out the analysis summary.
}
\details{
For class \link{cc} and \link{cc_ensemble}, the function will calculate confidence statistics for samples and clusters.
}
