% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{exclude_by_meta}
\alias{exclude_by_meta}
\title{Exclude Sequences from Alignment by Filtering on Metadata}
\usage{
exclude_by_meta(
  aln,
  field,
  pattern,
  retain = 0,
  seed = NULL,
  metadata = db_pull("metadata.Sequence.NCBI", silent = T)
)
}
\arguments{
\item{aln}{Alignment in DNAbin}

\item{field}{Character. Field (column) in metadata to apply the filter}

\item{pattern}{Character. Regex pattern to look for. Matching result will be excluded.}

\item{retain}{Numeric. Number of sequence to randomly exclude from exclusion (i.e. retain). Useful when excluding a group with large amount of sequence but keeping them represented in alignment. If bigger than number of sequences matching filtering criteria, no sequence will be excluded.}

\item{seed}{Will be passed to \link{set.seed}. Controls random selection of sequences when \code{retain} is not 0.}

\item{metadata}{A data frame. The metadata. Must incldue column given in \code{field}}
}
\value{
An alignment in DNAbin
}
\description{
Known issue: this function only filter the sequence based on single filtering criteria. For complex filter, take the source code and write for yourself.
}
