% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSD_handler.R
\name{extract_INSD_metadata}
\alias{extract_INSD_metadata}
\title{Extract INSD Metadata from \code{xml} File Downloaded From NCBI Genebank}
\usage{
extract_INSD_metadata(xml)
}
\arguments{
\item{xml}{File path to \code{xml} file}
}
\value{
A data.frame of flattened tabular metadata.
}
\description{
Extract INSD Metadata from \code{xml} File Downloaded From NCBI Genebank
}
