% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{fetch.land.elv}
\alias{fetch.land.elv}
\title{Fetch Land Elevation Data}
\usage{
fetch.land.elv(boundaries, crs = 4326, z = 3)
}
\arguments{
\item{boundaries}{A list of boundaries (min.lat, min.long, max.lat, max.long).}

\item{z}{Zoom level for elevation.}

\item{crrs}{Coorinate reference system.}
}
\description{
Fetch Land Elevation Data
}
