% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{fetch.NOAA.bathy}
\alias{fetch.NOAA.bathy}
\title{Fetch bathymetry map from NOAA.}
\usage{
fetch.NOAA.bathy(boundaries, resolution = 5, cache = T)
}
\arguments{
\item{cache}{Logical. Default to TRUE, the function will generate cache for the exact same boundary for future reuse.}
}
\description{
Fetch bathymetry map from NOAA.
}
