% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{fetch_land_elv}
\alias{fetch_land_elv}
\title{Fetch Land Elevation Data}
\usage{
fetch_land_elv(boundaries, crs = 4326, z = 3, cache = T)
}
\arguments{
\item{boundaries}{A list of boundaries (min.lat, min.long, max.lat, max.long).}

\item{z}{Zoom level for elevation.}

\item{cache}{Logical. Default to TRUE, the function will generate cache for the exact same boundary for future reuse.}
}
\description{
Fetch Land Elevation Data
}
