% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-formatting.R
\name{format_metadata.Specimen.Haploniscidae}
\alias{format_metadata.Specimen.Haploniscidae}
\title{Format Specimen Metadata}
\usage{
format_metadata.Specimen.Haploniscidae(tbl, formatting = T)
}
\description{
Format Specimen Metadata
}
