% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fasta_handler.R
\name{gblocks}
\alias{gblocks}
\title{Run Gblocks to Trim the Alignment}
\usage{
gblocks(
  fasta_path,
  gblocks_path = "/Users/hu_zhehao/Desktop/Biology-tools/Gblocks_0.91b/Gblocks",
  t = "d",
  b1 = "min",
  b2 = "min",
  b3 = 20,
  b4 = 2,
  b5 = "a",
  args = ""
)
}
\arguments{
\item{fasta_path}{File path of the alignment FASTA file.}

\item{gblocks_path}{File path of the Gblocks excutable.}

\item{t}{Type of sequence. Gblocks command line parameter, will be passed over as \code{-t}.}

\item{b1}{Minimum number of sequence for a conserved position (>= 50\%+1 of the number of sequences, default 50\%+1 of the number of sequences). Gblocks command line parameter, will be passed over as \code{-b1}. Also accepts \code{min} to choose the lowest value, which maximize the selected number of positions (i.e. least aggressive in trimming).}

\item{b2}{Minumum number of sequences for a flank position (>= b1, default 85\% of the number of sequences). Gblocks command line parameter, will be passed over as \code{-b2}. Also accepts \code{min} to choose the lowest value, which maximize the selected number of positions (i.e. least aggressive in trimming).}

\item{b3}{Maximum number of contiguous nonconserved positions (gblocks default to 8, here default to 20). Bigger values increase the selected number of positions.}

\item{b4}{Minimum length of a block after gap cleaning (gblocks default to 10, minimum value 2, here default to 2). Bigger values decrease the selected number of position.}

\item{b5}{Allowe gap positions. \code{n} for absolute no gap positions allowed in the final alignment. \code{h} for only positions where 50\% or more of the sequens have a gap are treated as gap position and are eliminated. \code{a} for all gap positions can be selected, no positions are eliminated becaused of the gaps.}

\item{args}{Other arguments that are passed directly to CLI command.}
}
\value{
Trimmed alignment will be written in new FASTA file under the same directory with suffix \verb{_gbtrimmed.fasta}. The path to the new file will be returned.
}
\description{
This is a wrapper function to call Gblocks in CLI. For details about the program please visit: https://home.cc.umanitoba.ca/~psgendb/doc/Castresana/Gblocks_documentation.html
}
