% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GIS.R
\name{geom_checkerboard}
\alias{geom_checkerboard}
\title{Add checkerboard to a ggplot}
\usage{
geom_checkerboard(
  min.lat,
  max.lat,
  lat.interval.count,
  min.long,
  max.long,
  long.interval.count,
  scale_only = F
)
}
\arguments{
\item{scale_only}{Logical. If TRUE no checkerboard will be drawn, but the plot will be scaled (cropped or expaneded) to the area according to giving boundaries.}
}
\value{
Updated ggplot objects.
}
\description{
Add checkerboard to a ggplot
}
