% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{gg_heatmap_hclust}
\alias{gg_heatmap_hclust}
\title{Generate Heatmap with Dendrogram in ggplot2 system.}
\usage{
gg_heatmap_hclust(
  hc.obj,
  matrix,
  heat_name = "Heat value",
  title = "",
  font_size = 8,
  dendro.stretch = c(0, 0.6),
  direction = 1,
  palette = NULL,
  zeroaswhite = T,
  grid = T,
  plot_labs = list(labs(x = NULL, y = NULL)),
  tip_length = 0,
  themes = list(theme()),
  heat_lab = NULL,
  args = list()
)
}
\arguments{
\item{hc.obj}{A hierarchical clustering object produced by \code{\link[stats:hclust]{stats::hclust()}}.}

\item{matrix}{The same matrix used for generating \code{hc.obj}, will be used for plotting the heatmap. The values must be normalized in range [0, 1] if \code{direction = 1}, or centralized at 0 with range [-1, 1] if \code{direction = 2}.}

\item{heat_name}{Character. Name of the parameter indicated by the heat, i.e. name of the values stored in the matrix.}

\item{title}{Character. Name of the plot.}

\item{font_size}{Numeric. Font size of the sample label.}

\item{dendro.stretch}{A numeric vector in length of 2. Controls vertical stretching of the dendrogram to align with heatmap. Will be passed to \code{expand} argument in \code{\link[ggplot2:scale_continuous]{ggplot2::scale_y_continuous()}}.}

\item{direction}{Numeric. Whether heatmap shows correlation in one direction (the level of correlation) or two (both the direction and the level of correlation).}

\item{palette}{A character vector of a color palette used for filling tiles in heatmap.}

\item{zeroaswhite}{Logical. Whether the 0 will be replaced by white.}

\item{grid}{Logical. Whether to show the grid in heatmap.}

\item{plot_labs}{\code{\link[ggplot2:labs]{ggplot2::labs()}} wrapped in a list, will be passed directly to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{tip_length}{Numeric. Length of elongation of the dendrogram tips.}

\item{themes}{\code{\link[ggplot2:theme]{ggplot2::theme()}} wrapped in a list, will be passed directly to \code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}.}

\item{heat_lab}{A character vector in length of 2. Labels to show in the legend for the minimal and maximal value.}

\item{args}{Further ggplot functions to be added to the plot.}
}
\value{
A ggplot object.
}
\description{
Generate Heatmap with Dendrogram in ggplot2 system.
}
\seealso{
\code{\link[stats:hclust]{stats::hclust()}}

\code{\link[ggplot2:ggplot]{ggplot2::ggplot()}}
}
