% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{gmyc_eval}
\alias{gmyc_eval}
\title{Perform GMYC on the Tree Generated by BEAST2}
\usage{
gmyc_eval(contree, prefix = NULL, method = "both")
}
\arguments{
\item{contree}{File path to the maximum credibility tree produced by BEAST2}

\item{prefix}{Character. Prefix to be added to the column name of the output data frame}

\item{method}{Character. Indicates whether to perform the single threshold (\code{single}) or the multiple threshold (\code{multiple}) GMYC, or \code{both}.}
}
\value{
A data frame of the delimitation result
}
\description{
Perform GMYC on the Tree Generated by BEAST2
}
\seealso{
\code{\link[=beast2_write_mrca]{beast2_write_mrca()}}
}
