% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi.R
\name{maldi_burn_metadata}
\alias{maldi_burn_metadata}
\title{Write Metadata into Spectra Object (burn the metadata in)}
\usage{
maldi_burn_metadata(
  rawSpectra = NULL,
  use_cache = F,
  metadata,
  maldi_voucher_col,
  metadata.map,
  inherit.name = T,
  write_cache = T,
  run.eval = NULL,
  cache_folder = NULL,
  write_metadata_field = T,
  return = T
)
}
\arguments{
\item{rawSpectra}{A MassSpectrum class object to be modified. Will be ignored if \code{use_cache} is set to \code{TRUE}.}

\item{use_cache}{Logical. Whether to load \code{rawSpectra} from the cache file. This could prevent from loading large object into your global environment. This option is only available when cache folder is specified by user configuration. See \code{\link[=create_config]{create_config()}}.}

\item{metadata}{A dataframe of the metadata.}

\item{maldi_voucher_col}{Column name in \code{metadata} that contains the unique voucher used as MALDI voucher (as in \code{\link[=maldi_set_eval_list]{maldi_set_eval_list()}})}

\item{inherit.name}{Logical. Whether to inherit the name of the metadata in \code{metadata} in the MassSpectrum object.}

\item{write_cache}{Logical. Whether to write cache file.}

\item{run.eval}{A dataframe produced by \code{\link[=maldi_set_eval_list]{maldi_set_eval_list()}}. If set to \code{NULL}, the function with default option will applied to \code{rawSpectra} to produce required dataframe.}

\item{cache_folder}{A path to cache folder. Default to \code{NULL} and the function will try to fetch the user configuration. If no path is set, the cache will be written to cache folder under current working directory.}

\item{write_metadata_field}{Logical. Whether to document the metadata fields written into the MassSpectrum object in this step. Must be \code{TRUE} if use \code{\link[=maldi_detect_peak]{maldi_detect_peak()}}.}

\item{return}{Logical. Wether to return the modified object.}

\item{metadata.extraction.list}{A list of one or two vectors of characters. The first vector should contain the columns name of the \code{metadata} that needs to be burnt into the MassSpectrum object. The second vector should contain the name of the metadata field burnt into the MassSpectrum object. Items from two vectors are paired by index. MALDI unique voucher specified in \code{maldi_voucher_col} should not be included and will be automatically added to the MassSpectrum object under the name \code{maldi_id}. If \code{inherit.name} is set to \code{TRUE} only one vector is required and the second vector will be ignored.}
}
\description{
Write Metadata into Spectra Object (burn the metadata in)
}
