% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi.R
\name{maldi_detect_peak}
\alias{maldi_detect_peak}
\title{Detect Peaks in Mass Spectra and Convert into Peak Matrix}
\usage{
maldi_detect_peak(
  averagedSpectra = NULL,
  use_cache = F,
  SNR = 8,
  hws = 8,
  method = "MAD",
  tolerance = 0.002,
  bin_method = "strict",
  min_peak = 25,
  metadata_field = NULL,
  lab_field = NULL,
  transform_method = "hellinger",
  write_cache = T,
  cache_folder = NULL,
  return = T
)
}
\description{
Detect Peaks in Mass Spectra and Convert into Peak Matrix
}
