% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi.R
\name{maldi_load_spectra_raw}
\alias{maldi_load_spectra_raw}
\title{Load Raw Bruker Flex Spectra}
\usage{
maldi_load_spectra_raw(raw_path = NULL, cache_folder = NULL, return = T)
}
\arguments{
\item{raw_path}{Path to folder containing subfolder of each measurement.}

\item{cache_folder}{Path to output cache. Default to \code{NULL}. If set to \code{NULL}, it will try to use user configuration (see \code{create_config}), if user configuration is not set, it will create a 'cache/' folder under current working directory.}

\item{return}{Logical. Whether to return the imported MassSpectrum object.}
}
\description{
\code{maldi_load_spectra_raw} loads raw MALDI-TOF Spectra from Bruker Flex files or cache if possible.
}
\details{
MALDI-TOF Raw Spectra are read internally by \code{MALDIquantForeign::importBrukerFlex(removeEmptySpectra = TRUE)}
}
