% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi.R
\name{maldi_process_spectra}
\alias{maldi_process_spectra}
\title{Process Mass Spectra for Normalized, Baseline Removed Spectra Ready for Comparison}
\usage{
maldi_process_spectra(
  rawSpectra.rich = NULL,
  use_cache = F,
  report = F,
  report_folder = NULL,
  trim.min = 2000,
  trim.max = 20000,
  transform_method = "sqrt",
  smooth_method = "SavitzkyGolay",
  smooth_hws = 10,
  baseline_method = "SNIP",
  SNIP_iteration = 20,
  calibrate_method = "TIC",
  average_method = "mean",
  pause = F,
  write_cache = T,
  cache_folder = NULL,
  return = T
)
}
\arguments{
\item{rawSpectra.rich}{A MassSpectrum class object with burnt-in metadata. Will be ignored if \code{use_cache} is set to \code{TRUE}.}

\item{use_cache}{Logical. Whether to load \code{rawSpectra.rich} from the cache file. This could prevent from loading large object into your global environment. This option is only available when cache folder is specified by user configuration. See \code{\link[=create_config]{create_config()}}.}

\item{report}{Logical. Whether to produce pdf report of the spectra for each processing step.}

\item{report_folder}{A path to a folder to export the pdf report. If set to \code{NULL}, the function will try to fetch the path from user configuration.}

\item{trim.min}{Numeric. Minimal mass to be included. Unit in Da.}

\item{trim.max}{Numeric. Maximal mass to be included. Unit in Da.}

\item{transform_method}{Character. Method for intensity transformation on the mass spectrum. Will be passed to \code{\link[MALDIquant:transformIntensity-methods]{MALDIquant::transformIntensity()}}.}

\item{smooth_method}{Character. Method for smoothing mass spectrum. Will be passed to \code{\link[MALDIquant:smoothIntensity-methods]{MALDIquant::smoothIntensity()}}}

\item{smooth_hws}{Numerical. Half window size applied to mass spectrum smoothing. Will be passed to \code{\link[MALDIquant:smoothIntensity-methods]{MALDIquant::smoothIntensity()}}}

\item{baseline_method}{Character. Method for baseline removing of the mass spectrum. Will be passed to \code{\link[MALDIquant:removeBaseline-methods]{MALDIquant::removeBaseline()}}}

\item{SNIP_iteration}{Numerical. Iterations to be used in baseline removal with SNIP algorithm. Details see \code{\link[MALDIquant:estimateBaseline-methods]{MALDIquant::estimateBaseline()}}. Will be passed to \code{\link[MALDIquant:removeBaseline-methods]{MALDIquant::removeBaseline()}}}

\item{calibrate_method}{Character. Method used for mass spectrum intensity calibration. Will be passed to \code{\link[MALDIquant:calibrateIntensity-methods]{MALDIquant::calibrateIntensity()}}}

\item{average_method}{Character. Method used for calculating average spectrum from each sample. Will be passed to \code{\link[MALDIquant:averageMassSpectra-functions]{MALDIquant::averageMassSpectra()}}}

\item{pause}{Logical. Whether to pause after every report is generated. This enables to stop the analysis in the middle to adjust parameters.}

\item{write_cache}{Logical. Whether to write cache file.}

\item{cache_folder}{A path to cache folder. Default to \code{NULL} and the function will try to fetch the user configuration. If no path is set, the cache will be written to cache folder under current working directory.}

\item{return}{Logical. Wether to return the modified object.}
}
\description{
Process Mass Spectra for Normalized, Baseline Removed Spectra Ready for Comparison
}
