% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maldi.R
\name{maldi_set_eval_list}
\alias{maldi_set_eval_list}
\title{Measurement Version Management}
\usage{
maldi_set_eval_list(rawSpectra, last_attempt = T, sep = "_")
}
\arguments{
\item{rawSpectra}{Raw Spectra.}

\item{last_attempt}{Logical. Whether evaluate the last attempt of all samples. Interactive step will be introduced if set to \code{FALSE}.}

\item{sep}{Seperator used to split the measurement name into voucher and attempt. See details.}
}
\value{
A dataframe with four columns: \describe{
\item{\code{maldi_voucher}: sample voucher}
\item{\code{total_attemt}: Total measurements of the sample}
\item{\code{eval}: Which measurement to be evaluated}
\item{\code{run_name}: Measrument run name to be evaluated}
}
}
\description{
Measurement Version Management
}
\details{
\code{maldi_set_eval_list} relies on custom naming convention that named the MALDI-TOF measurments under the rule \code{<MALDI-TOF ID>_<attempt>} (e.g. VPSM001-1, VPSM003-3), which enables the function to effectively identify the attempts on the same sample and provide an interactive step for manually choosing attempts that are not the last attempt of the same samples to be evaluated. Default workflow always evaluates the last (non-empty) spectrum.
}
