% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{map_values}
\alias{map_values}
\title{Map a vector's values using a lookup table}
\usage{
map_values(input_vector, lut_df, key_col, value_col)
}
\arguments{
\item{input_vector}{The vector of original values}

\item{lut_df}{The lookup data frame}

\item{key_col}{The unquoted column name in \code{lut_df} to match against.}

\item{value_col}{The unquoted column name in \code{lut_df} to get the new values from.}
}
\value{
A new vector with the mapped values, in the same order as the input.
}
\description{
Map a vector's values using a lookup table
}
