% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/consensus_clustering.R
\name{md_report}
\alias{md_report}
\alias{md_report-ccmk}
\alias{md_report-cc_ensemble}
\title{Create Markdown Report}
\usage{
md_report(x, ...)
}
\arguments{
\item{report_path}{Path to export the generated figures.}

\item{report_name}{A run name to put on the report and file name.}
}
\value{
A list of markdown content used for producing the pdf report.
}
\description{
Create Markdown Report

Generate Markdown Report for \code{ccmk} Object

Generate Markdown Report for \code{cc_ensemble} Object
}
