% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes_helper.R
\name{mrbayes}
\alias{mrbayes}
\alias{mrbayes_mcmc_cmd}
\alias{mrbayes_partition_cmd}
\alias{mrbayes_lset_cmd}
\alias{mrbayes_prset_cmd}
\alias{mrbayes_sum_cmd}
\alias{mrbayes_block}
\alias{mrbayes-args}
\title{Helper Functions to Generate MrBayes Block in Nexus}
\usage{
mrbayes_mcmc_cmd(gen = 1e+07, nchains = 4, nruns = 4, stopval = 0.01, ...)

mrbayes_partition_cmd(loci_length, loci_name, protein_coding_loci)

mrbayes_lset_cmd(nst, rates, ..., applyto = NULL)

mrbayes_prset_cmd(..., applyto = NULL)

mrbayes_sum_cmd(..., cmd)

mrbayes_block(...)
}
\arguments{
\item{gen}{Number of MCMC generations}

\item{nchains}{Number of MCMC chains}

\item{nruns}{Number of parallel runs}

\item{stopval}{Threshold value to stop the run}

\item{...}{Additional arguments. \describe{
\item{\link{mrbayes_mcmc_cmd}}{Further parameters to put in mcmc command}
\item{\link{mrbayes_lset_cmd}}{Further parameters to put in lset command}
\item{\link{mrbayes_prset_cmd}}{Further parameters to put in prset command}
\item{\link{mrbayes_sum_cmd}}{Further parameters to put in sumt or sump command, when given, \code{cmd} must also be provided to specify whether sumt or sump is being generated}
\item{\link{mrbayes_block}}{Commands to be included in the MrBayes block}
}}

\item{loci_length}{A numeric vector. Length of each locus.}

\item{loci_name}{A character vector. Name of each locus.}

\item{protein_coding_loci}{A character vector of the names of the protein coding locus. The locus will be partitioned further based on codon position to allow different substitution rate at each position.}

\item{nst}{Value passed to \code{lset nst}}

\item{rates}{Value passed to \code{lset rates}}

\item{applyto}{A numeric vector, indicating the index of the partition to apply the \code{lset} settings to. Default to \code{NULL}, which applies to the whole alignment.}

\item{cmd}{Either \code{"sump"} or \code{"sumt"}, indicating the generating command. Will be ignored when \code{...} is missing.}
}
\value{
A character string.
}
\description{
These functions help generate MrBayes command to put in nexus block that can be written with the alignment together. \code{mrbayes_mcmc_cmd}, \code{mrbayes_partition_cmd}, \code{mrbayes_lset_cmd}, \code{mrbayes_prset_cmd}, \code{mrbayes_sum_cmd} produce individual part of the command. Their output can be passed into \code{mrbayes_block} to package them as a complete nexus block. Also see \code{\link[=prepare_mrbayes_nexus]{prepare_mrbayes_nexus()}} for a more powerful wrapper.
}
\seealso{
\code{\link[=prepare_mrbayes_nexus]{prepare_mrbayes_nexus()}}
}
