% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mrbayes_helper.R
\name{prepare_mrbayes_nexus}
\alias{prepare_mrbayes_nexus}
\title{Helper-Wrapper for Generating Nexus File for MrBayes}
\usage{
prepare_mrbayes_nexus(
  alignment,
  output_path,
  mcmc_options,
  loci_length,
  loci_name,
  protein_coding_loci,
  lset_options,
  prset_options = NULL,
  sum_options = NULL
)
}
\arguments{
\item{alignment}{Alignment in DNAbin}

\item{output_path}{Path of the output nexus file}

\item{mcmc_options}{A named list of arguments to be passed to \code{\link[=mrbayes_mcmc_cmd]{mrbayes_mcmc_cmd()}}}

\item{loci_length}{A numeric vector. Length of each locus. Will be passed to \code{\link[=mrbayes_partition_cmd]{mrbayes_partition_cmd()}}}

\item{loci_name}{A character vector. Name of each locus. Will be passed to \code{\link[=mrbayes_partition_cmd]{mrbayes_partition_cmd()}}}

\item{protein_coding_loci}{A character vector of the names of the protein coding locus. The locus will be partitioned further based on codon position to allow different substitution rate at each position. Will be passed to \code{\link[=mrbayes_partition_cmd]{mrbayes_partition_cmd()}}}

\item{lset_options}{A named list of arguments to be passed to \code{\link[=mrbayes_lset_cmd]{mrbayes_lset_cmd()}}}

\item{sum_options}{A named list of arguments to be passed to \code{\link[=mrbayes_prset_cmd]{mrbayes_prset_cmd()}}}

\item{prset_option}{A named list of arguments to be passed to \code{\link[=mrbayes_prset_cmd]{mrbayes_prset_cmd()}}}
}
\value{
The file path of written nexus file.
}
\description{
Takes arguments for MrBayes helper construct nexus file with mrbayes block, and export.
}
\seealso{
\link{mrbayes}
}
