% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{read_contree}
\alias{read_contree}
\title{Read MrBayes Output Consensus Tree}
\usage{
read_contree(assets_date, subfolder)
}
\arguments{
\item{assets_date}{For locating asset folder.}

\item{subfolder}{Subfolder path within the daily log asset folder.}
}
\value{
A tidytree treedata object.
}
\description{
Read MrBayes Output Consensus Tree
}
