% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{read_contree_mrbayes}
\alias{read_contree_mrbayes}
\title{Read MrBayes Output Consensus Tree}
\usage{
read_contree_mrbayes(assets_path)
}
\arguments{
\item{assets_path}{Path to \code{.con.tre} file or the folder containing the file. If to the folder, there could only be one \code{.con.tre} file in the folder.}
}
\value{
A tidytree treedata object.
}
\description{
Read MrBayes Output Consensus Tree
}
