% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/quarto.R
\name{render_cc_report}
\alias{render_cc_report}
\title{Wrapper for Quarto Render Consensue Clustering Result Report}
\usage{
render_cc_report()
}
\description{
Wrapper for Quarto Render Consensue Clustering Result Report
}
