% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R, R/consensus_clustering.R
\name{report}
\alias{report}
\alias{report-cc_eval}
\title{Report Analysis Result}
\usage{
report(x, ...)

report(cc_eval)
}
\description{
A S7 generics to print out the analysis summary.
}
