% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{split_concat_alignment}
\alias{split_concat_alignment}
\title{Split Concatenated Alignment Back into Single-locus Alignment}
\usage{
split_concat_alignment(alignment, loci_name, loci_length, output_folder)
}
\arguments{
\item{alignment}{Concatenated alignment}

\item{loci_name}{Names of the loci}

\item{loci_length}{Lengths of each loci}

\item{output_folder}{Output folder}
}
\description{
Why should I do this? You might ask. Well, to get the actual sequences of each locus used in subsequent analysis to perform the model test.
}
\seealso{
\code{\link[=concatenate_alignment]{concatenate_alignment()}}
}
