% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{test_sub_model}
\alias{test_sub_model}
\title{Test for Best Fit Model in IQ-TREE}
\usage{
test_sub_model(
  fasta_path,
  run_name,
  output_folder,
  iqtree_exe =
    "/Users/hu_zhehao/Desktop/Biology-tools/iqtree-2.3.6-macOS-arm/bin/iqtree2"
)
}
\arguments{
\item{fasta_path}{Path to fasta file of the alignment}

\item{run_name}{Character. A custom name to be added to output files}

\item{output_folder}{Path to generate all output files from IQ-TREE}

\item{iqtree_exe}{Path to the IQ-TREE excutable}
}
\value{
Path to the output folder
}
\description{
This function calls local IQ-TREE CLI excutable to perform the modelFinder.
}
