% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{tree_eval}
\alias{tree_eval}
\title{Wrapper for Full Consensus Tree Visualization Process}
\usage{
tree_eval(
  assets_date,
  subfolder = "",
  outgroup,
  outgroups = NULL,
  xlim.factor = 1.5,
  width = 15,
  height = 30
)
}
\arguments{
\item{assets_date}{For locating asset folder.}

\item{subfolder}{Subfolder path within the daily log asset folder.}

\item{outgroup}{Name of the outgroup that can be used to find all outgroup taxa. The name will be matched against tip label of the tree, and the best matching labels are used as outgroup. This argument will be ignored if \code{outgroups} is given.}

\item{outgroups}{A vector of characters that contains all taxa labels of the outgroups.}

\item{xlim.factor}{A factor to expand the x axis.}

\item{width}{SVG width.}

\item{height}{SVG height.}
}
\value{
The path leading to written svg file of the tree.
}
\description{
Wrapper for Full Consensus Tree Visualization Process
}
