% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{tree_reroot}
\alias{tree_reroot}
\title{Reroot Consensus Tree Produced by MrBayes}
\usage{
tree_reroot(tree, outgroup, outgroups = NULL)
}
\arguments{
\item{tree}{A tidytree treedata object.}

\item{outgroup}{Name of the outgroup that can be used to find all outgroup taxa. The name will be matched against tip label of the tree, and the best matching labels are used as outgroup. This argument will be ignored if \code{outgroups} is given.}

\item{outgroups}{A vector of characters that contains all taxa labels of the outgroups.}
}
\value{
A rerooted tidytree treedata object.
}
\description{
Reroot Consensus Tree Produced by MrBayes
}
\details{
A proper rerooting is currently only guaranteed for consensus tree produced by MrBayes, due to inconsistent usage of the branch support value across variety of phylogenetic analysis programs. Details see Czech et al. 2017.
}
\references{
Czech, L., Huerta-Cepas, J. and Stamatakis, A. (2017) “A Critical Review on the Use of Support Values in Tree Viewers and Bioinformatics Toolkits,” Molecular Biology and Evolution, 34(6), pp. 1535–1542. Available at: https://doi.org/10.1093/molbev/msx055.
}
