% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{tree_visualize}
\alias{tree_visualize}
\title{Visualize Phylogenetic Tree}
\usage{
tree_visualize(tree, xlim.factor = 1.5, align = F)
}
\arguments{
\item{tree}{A tree for visualization.}

\item{xlim.factor}{A factor to expand the x axis.}
}
\value{
A ggtree object.
}
\description{
Visualize Phylogenetic Tree
}
