% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/INSD_handler.R
\name{update_INSD_metadata}
\alias{update_INSD_metadata}
\title{Wrapper Function to Read All \code{XML} File and Update Information to Databse.}
\usage{
update_INSD_metadata(XML_folder = "data/sequence/INSD", regenerate = F)
}
\arguments{
\item{XML_folder}{Path to folder containing \code{xml} files. All \code{xml} files in the folder will be loaded.}

\item{regenerate}{logical. Whether to overwrite the table in database (i.e. regenerating the table). Default to FALSE, which will only add metadata that has not been added in the table.}
}
\value{
Nothing.
}
\description{
Wrapper Function to Read All \code{XML} File and Update Information to Databse.
}
