% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence_mapper.R
\name{update_sequence_map}
\alias{update_sequence_map}
\title{Create a Summary Table Mapping Sequence Label/Accession Number with Vouchered Animals.}
\usage{
update_sequence_map(regenerate = F)
}
\arguments{
\item{regenerate}{logical. Whether to overwrite existing table in the database. Otherwise only the rows with vouchered animal not yet in the database will be added.}
}
\value{
Nothing. A data.frame will be written into the database.
}
\description{
This will summarize the sequence labels in the way that the sequnces originated from the same vouchered animals could be linked together, which is essential for concatenate multilocus alignment.
}
