% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{visualize_tree}
\alias{visualize_tree}
\title{Visualize Consensus Tree Produced From MrBayes Analysis}
\usage{
visualize_tree(assets_date, outgroup, subfolder = "", xlim.factor = 1.5)
}
\arguments{
\item{assets_date}{Date of the asset folder where the MrBayes analysis process is located in. This is a project-specific data structure adaptation. It just simplifies the location of the target files.}

\item{outgroup}{Sequence labels of the outgroup sequences. It will use automatic gap detection to find the sequences so it allows a certain degree of ambiguity in the name.}

\item{subfolder}{Subfolder where the MrBayes process is located under the asset folder of the date. This is a project-specific data structure adaptation. It just simplifies the location of the target files.}

\item{xlim.factor}{Factor to expand the x-axis of the tree. Increase this value if long labels are omitted because they exceed the plotting area.}
}
\value{
A ggplot of visualized tree.
}
\description{
Visualize Consensus Tree Produced From MrBayes Analysis
}
