% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sequence.R
\name{clean_label}
\alias{clean_label}
\title{Clean Sequence Labels in Alignment}
\usage{
clean_label(fasta_path)
}
\arguments{
\item{fasta_path}{Path to FASTA file}
}
\value{
A new FASTA file will be written in the same directory. The path to the new file will be returned.
}
\description{
This only applies to DZMB collection where the sequences are managed using field ID. The gene will be determined by the file name (whether it's COI, 18S, or 28S) and a gene suffix will be added to the voucher (e.g. VPS101_COI, ZHH003_18S).
}
