% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tbl-formatting.R
\name{format_metadata.Sequence.NCBI}
\alias{format_metadata.Sequence.NCBI}
\title{Format NCBI Sequence Metadata}
\usage{
format_metadata.Sequence.NCBI(tbl, formatting = T, format = 1)
}
\description{
Format NCBI Sequence Metadata
}
