% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{tree_relabel}
\alias{tree_relabel}
\title{Relabel Tree's Tip Labels}
\usage{
tree_relabel(tree)
}
\arguments{
\item{tree}{A tidytree treedata object.}
}
\value{
A relabeled tidytree treedata object.
}
\description{
Relabel Tree's Tip Labels
}
