% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tree_handler.R
\name{tree_visualize}
\alias{tree_visualize}
\title{Visualize Phylogenetic Tree}
\usage{

tree_visualize(
  tree,
  xlim.factor = 1.5,
  align = F,
  taxa.font.size = 3.5,
  format.taxa = T,
  linesize = NULL
)
}
\arguments{
\item{tree}{A tree for visualization.}

\item{xlim.factor}{A factor to expand the x axis.}

\item{linesize}{Thickness of the tree lines. Default to \code{NULL}, which is set automatically by the function.}
}
\value{
A ggtree object.
}
\description{
Visualize Phylogenetic Tree
}
